pyproject.toml
[project]
name = "foxholed"
version = "0.1.0"
description = "Foxhole game map position reader"
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.6",
    "opencv-python-headless>=4.9",
    "pytesseract>=0.3",
    "mss>=9.0",
    "numpy>=1.26",
    "Pillow>=10.0",
]

[project.scripts]
foxholed = "foxholed.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/foxholed"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
